tags/v1.5.2/marmot/requirements.txt
Django==1.8.13
django-guardian==1.4.4
django-fsm==2.4.0
MySQL-python==1.2.5
redis==2.10.5
zeroc-ice==3.6.2.1